Refactor click tracking
