+++ src/app/containers/MostRead/Canonical/index.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState, useContext } from 'react';
 import 'isomorphic-fetch';
-import { oneOf, string, elementType } from 'prop-types';
+import { oneOf, string, elementType, shape } from 'prop-types';
 import { MostReadLink, MostReadItemWrapper } from './Item';
 import MostReadList from './List';
 import MostReadRank from './Rank';
@@ -16,54 +16,6 @@ import {
   MOST_READ_FETCH_ERROR,
 } from '#lib/logger.const';
 import useViewTracker from '#hooks/useViewTracker';
-import useClickTrackerHandler from '#hooks/useClickTrackerHandler';
-
-const TrackedMostReadLink = ({
-  dir,
-  service,
-  script,
-  title,
-  href,
-  size,
-  eventTrackingData,
-  children,
-}) => {
-  const clickTrackerHandler = useClickTrackerHandler(eventTrackingData);
-
-  return (
-    <MostReadLink
-      dir={dir}
-      service={service}
-      script={script}
-      title={title}
-      href={href}
-      size={size}
-      onClick={clickTrackerHandler}
-    >
-      {children}
-    </MostReadLink>
-  );
-};
-
-TrackedMostReadLink.propTypes = {
-  dir: oneOf(['rtl', 'ltr']),
-  service: string.isRequired,
-  script: shape(scriptPropType).isRequired,
-  title: string.isRequired,
-  href: string.isRequired,
-  children: node, // this node will be a timestamp container
-  size: oneOf(['default', 'small']),
-  eventTrackingData: shape({
-    componentName: string,
-  }),
-};
-
-TrackedMostReadLink.defaultProps = {
-  dir: 'ltr',
-  children: null,
-  size: 'default',
-  eventTrackingData: null,
-};
 
 const logger = nodeLogger(__filename);
 
@@ -171,7 +123,7 @@ const CanonicalMostRead = ({
               columnLayout={columnLayout}
               size={size}
             />
-            <TrackedMostReadLink
+            <MostReadLink
               dir={dir}
               service={service}
               script={script}
@@ -190,7 +142,7 @@ const CanonicalMostRead = ({
                   timezone={timezone}
                 />
               )}
-            </TrackedMostReadLink>
+            </MostReadLink>
           </MostReadItemWrapper>
         ))}
       </MostReadList>
